Changed MaxDuration to float

+++ Data/Scripts/WeaponThread/script/WeaponStructure.cs
@@ -656,7 +656,7 @@ namespace WeaponThread
             {
                 [ProtoMember(1)] internal float Scale;
                 [ProtoMember(2)] internal float MaxDistance;
-                [ProtoMember(3)] internal uint MaxDuration;
+                [ProtoMember(3)] internal float MaxDuration;
                 [ProtoMember(4)] internal bool Loop;
                 [ProtoMember(5)] internal bool Restart;
                 [ProtoMember(6)] internal float HitPlayChance;
